remove and abstract Puppet dependency
This removes the hard dependency on the `puppet` gem. It will simply use
whichever gem you have installed that provides the `puppet` namespace.

Con: this means that it will no longer automatically install all of its
     dependencies. You must do it yourself.

It also sets the stage for abstracting out the telemetry backend. If
`Puppet.implementation` returns 'openvox' then it will load an alternate
backend, which is currently a noop because we don't have one stood up yet.

Warning: this will silently break both our and Perforce's server
packaging job. An appropriate implementation gem must be added to
https://github.com/OpenVoxProject/openvox-server/blob/main/resources/ext/build-scripts/dropsonde-gem.txt

(or we can do a better packaging job, lol)